ExportTool 1.0

+++ unity-plugin/core/editor/src/base/ExportStore.cs
@@ -1,6 +1,8 @@
 ﻿using System.Collections.Generic;
 using System.IO;
 using System.Text;
+using ICSharpCode.SharpZipLib.Core;
+using ICSharpCode.SharpZipLib.Zip;
 using UnityEditor;
 using UnityEngine;
 
@@ -12,12 +14,16 @@ namespace WeChat
         static BeefBallConfig globalExportConfig = null;
         private static void InitGlobalExportConfig()
         {
+            if (globalExportConfig != null)
+            {
+                return;
+            }
             globalExportConfig = DirectoryUtil.GetScriptableObject<BeefBallConfig>("beefball");
             if (globalExportConfig == null)
             {
                 globalExportConfig = DirectoryUtil.CreateScriptableObject<BeefBallConfig>("beefball");
             }
-            
+
             // Debug.Log(globalExportConfig);
             if (globalExportConfig != null && globalExportConfig.storagePath != null && globalExportConfig.storagePath != "")
             {
@@ -25,7 +31,6 @@ namespace WeChat
             }
             // Debug.Log(2222);
         }
-
 
         static ExportStore()
         {
@@ -71,6 +76,11 @@ namespace WeChat
         {
             get
             {
+                if (!globalExportConfig)
+                {
+                    Debug.LogError("请先选择导出路径:导出/手动选择导出路径");
+                    return "";
+                }
                 return globalExportConfig.exportPath;
             }
             set
@@ -85,6 +95,11 @@ namespace WeChat
         {
             get
             {
+                if (!globalExportConfig)
+                {
+                    Debug.LogError("请先选择导出路径:导出/手动选择导出路径");
+                    return "";
+                }
                 return globalExportConfig.storagePath;
             }
             set
@@ -108,7 +123,6 @@ namespace WeChat
                 }
             }
         }
-
 
         // 初始化目标目录，包括storage.json和两个存储目录
         public static void init()
@@ -178,6 +192,10 @@ namespace WeChat
 
         private static bool CheckCacheFileCompletion()
         {
+            if (storage == null)
+            {
+                return false;
+            }
             List<string> assetsMD5 = new List<string>();
             JSONObject files = storage.GetField("files");
             for (int i = 0; i < files.Count; i++)
@@ -205,6 +223,10 @@ namespace WeChat
 
         public static void ResetStorage()
         {
+            if (GetExportStoragePath() == "")
+            {
+                return;
+            }
             if (File.Exists(GetExportStoragePath()))
             {
                 File.Delete(GetExportStoragePath());
@@ -360,7 +382,9 @@ namespace WeChat
             SaveFileManifest(relativePath, fileType, assetMD5);
         }
 
-
+        /*
+         * 往缓存库里添加一个资源
+         */
         public static void AddResource(
             string descriptionFilePath,
             string resourceType,
@@ -369,7 +393,9 @@ namespace WeChat
             List<string> dependencies,
 
             // 这个资源使用到的文件
-            List<string> useFile
+            List<string> useFile,
+
+            JSONObject importSetting = null
         )
         {
             if (descriptionFilePath == null || descriptionFilePath == "")
@@ -384,6 +410,11 @@ namespace WeChat
             fileItem.AddField("type", resourceType);
             fileItem.AddField("dependencies", WXUtility.ConvertListToJSONArray(dependencies));
             fileItem.AddField("useFile", WXUtility.ConvertListToJSONArray(useFile));
+            if (importSetting != null)
+            {
+                fileItem.AddField("importSetting", importSetting);
+            }
+
             storageResourceJSON.AddField(descriptionFilePath, fileItem);
 
             saveStorage();
@@ -391,7 +422,7 @@ namespace WeChat
 
         public static void GenerateResourcePackage(string packageName, string resourcePath, string packageTypeName = ".mgepackage/")
         {
-            string destPath = Path.Combine(exportPath == "" ?  ExportStore.storagePath : exportPath, packageName + packageTypeName);
+            string destPath = Path.Combine(exportPath == "" ? ExportStore.storagePath : exportPath, packageName + packageTypeName);
 
             List<string> resources = new List<string>();
             resources.Add(resourcePath);
@@ -401,13 +432,11 @@ namespace WeChat
         }
         public static void GenerateResourcePackage(string packageName, List<string> resources, string packageTypeName = ".mgepackage/")
         {
-            string destPath = Path.Combine(exportPath == "" ?  ExportStore.storagePath : exportPath, packageName + packageTypeName);
+            string destPath = Path.Combine(exportPath == "" ? ExportStore.storagePath : exportPath, packageName + packageTypeName);
 
             WXExportPackageGenerator generator = new WXExportPackageGenerator(destPath, resources);
             generator.Generate();
         }
-
-
 
         class WXExportPackageGenerator
         {
@@ -435,20 +464,33 @@ namespace WeChat
             }
 
             /**
-            * 生成一个资源包
-            * @param destPath 目标路径
-            * @param resources 要打进去的资源路径
-            */
+             * 生成一个资源包
+             * @param destPath 目标路径
+             * @param resources 要打进去的资源路径
+             */
             public void Generate()
             {
+                // string zipPath = destPath[destPath.Length - 1] == '/' ? destPath.Substring (0, destPath.Length - 1) : destPath;
+
+                // using (FileStream fsOut = File.Create (zipPath))
+                // using (ZipOutputStream zipStream = new ZipOutputStream (fsOut)) {
+                //     foreach (string resourceDescriptionFilePath in entryResources) {
+                //         writeResourceRecursive (resourceDescriptionFilePath, zipStream);
+                //     }
+
+                //     zipStream.PutNextEntry (new ZipEntry ("group.manifest.json"));
+                //     byte[] groupManifestBytes = Encoding.UTF8.GetBytes (packageGroupManifest.Print (true));
+                //     zipStream.Write (groupManifestBytes, 0, groupManifestBytes.Length);
+                //     zipStream.CloseEntry ();
+                //     zipStream.Close ();
+                // }
                 foreach (string resourceDescriptionFilePath in entryResources)
                 {
                     writeResourceRecursive(resourceDescriptionFilePath);
                 }
-
-                // 写入group.manifest.json文件
                 wxFileUtil.SaveJsonFile(packageGroupManifest, Path.Combine(destPath, "group.manifest.json"));
             }
+
 
             private void writeResourceRecursive(string resourceDescriptionFilePath)
             {
@@ -473,6 +515,10 @@ namespace WeChat
                 definitionObject.AddField("dependencies", resourceStorage.GetField("dependencies"));
                 definitionObject.AddField("type", resourceStorage.GetField("type"));
                 definitionObject.AddField("descriptionFileID", resourceDescriptionFilePath);
+                if (resourceStorage.GetField("importSetting") != null)
+                {
+                    definitionObject.AddField("importSetting", resourceStorage.GetField("importSetting"));
+                }
 
                 // 把记在缓存里的useFile拿出来，遍历，写入group.manifest.json的fileDescription字段
                 // 并且排除掉resourceDescriptionFilePath之后放入resourceDefinition
@@ -516,6 +562,77 @@ namespace WeChat
                 }
             }
 
+            private void writeResourceRecursive(string resourceDescriptionFilePath, ZipOutputStream zipStream)
+            {
+                // 如果已经转换过了
+                if (packageGroupManifest_resourceDefinitions.HasField(resourceDescriptionFilePath))
+                {
+                    return;
+                }
+                // 先找到之前已经写入缓存表的资源。
+                JSONObject resourceStorage = storage.GetField("assets").GetField(resourceDescriptionFilePath);
+                if (resourceStorage == null)
+                {
+                    Debug.LogError("创建资源包时写入'" + resourceDescriptionFilePath + "'失败，没有找到该资源的转换记录");
+                    return;
+                }
+
+                // 创建每一个资源definition的json对象
+                JSONObject definitionObject = new JSONObject(JSONObject.Type.OBJECT);
+                packageGroupManifest_resourceDefinitions.AddField(resourceDescriptionFilePath, definitionObject);
+
+                // 写入字段
+                definitionObject.AddField("dependencies", resourceStorage.GetField("dependencies"));
+                definitionObject.AddField("type", resourceStorage.GetField("type"));
+                definitionObject.AddField("descriptionFileID", resourceDescriptionFilePath);
+                if (resourceStorage.GetField("importSetting") != null)
+                {
+                    definitionObject.AddField("importSetting", resourceStorage.GetField("importSetting"));
+                }
+
+                // 把记在缓存里的useFile拿出来，遍历，写入group.manifest.json的fileDescription字段
+                // 并且排除掉resourceDescriptionFilePath之后放入resourceDefinition
+                JSONObject outputUseFileArray = new JSONObject(JSONObject.Type.ARRAY);
+                definitionObject.AddField("useFile", outputUseFileArray);
+
+                // 放入文件
+                foreach (string usingFile in WXUtility.ConvertJSONArrayToList(resourceStorage.GetField("useFile")))
+                {
+                    if (usingFile != resourceDescriptionFilePath)
+                    {
+                        outputUseFileArray.Add(usingFile);
+                    }
+                    if (packageGroupManifest_fileDescription.HasField(usingFile))
+                    {
+                        continue;
+                    }
+                    JSONObject fileStorage = storage.GetField("files").GetField(usingFile);
+
+                    // add fileDescriptions
+                    JSONObject fileDescription = new JSONObject(JSONObject.Type.OBJECT);
+                    packageGroupManifest_fileDescription.AddField(usingFile, fileDescription);
+                    fileDescription.AddField("path", usingFile);
+
+                    // add files 555
+                    JSONObject fileItem = new JSONObject(JSONObject.Type.OBJECT);
+                    fileItem.AddField("path", usingFile);
+                    fileItem.AddField("filetype", fileStorage.GetField("filetype"));
+                    packageGroupManifest_files.Add(fileItem);
+
+                    zipStream.PutNextEntry(new ZipEntry(usingFile));
+                    var buffer = new byte[10240];
+                    using (FileStream fsInput = File.OpenRead(Path.Combine(storagePath, Path.Combine(CONTENT_FOLDER, fileStorage.GetField("MD5").GetRawString()))))
+                    {
+                        StreamUtils.Copy(fsInput, zipStream, buffer);
+                    }
+                }
+
+                // 递归转依赖
+                foreach (string dependencyResource in WXUtility.ConvertJSONArrayToList(resourceStorage.GetField("dependencies")))
+                {
+                    writeResourceRecursive(dependencyResource, zipStream);
+                }
+            }
 
             /**
              * 递归获取一段资源列表的所有依赖资源
